Add files via upload
This project aims at solving a project question which involves finding the average Income of individuals who purchased and did not purchase a bike since I worked on a bike buyer dataset, the distance of the bike and the age range of those who are purchasing the bike.

The first procedure was to create a new working sheet which I used in my analysis procedure by copying the dataset from the raw data into the new working sheet as I was taught not to work directly on the raw data given.
Afterwards, I had to ensure that my dataset is well formatted, removed the duplicates and I had to highlight the dataset I will be using in answering my project questions.
I had to replace the Marital Status and Gender column with the appropriate designation to prevent confusion while plotting my chart. I replaced the M and S in the Marital Status with Married and Single and the M and F in the Gender with Male and Female.
I carried out an IF statement in the Age bracket column to ensure we only have the age bracket of Adolescent, Middle age and old age.
I later created my Pivot Table for the average income for the two genders, the average miles, and the age range and I created my chart from the three Pivot Table created. I also labelled the charts and edited all that needs to be edited.
I created my dashboard and named it Bike Sales Dashboard, created a slicer for the marital status, region, and education and ensured that the three Pivot Tables are connected together using the report connection tool.